EventSelector: Smaller bugfixes in electron selection.
- Two variable that were meant to be ints were defined as doubles.

- 'Reject out of time' was only asked for barrel electrons.

+++ Main/EventSelector.hh
@@ -171,7 +171,7 @@ private:
    double const m_ele_barrel_HcalD1_offset;
    double const m_ele_barrel_HcalD1_slope;
    double const m_ele_barrel_HcalD1_rhoSlope;
-   double const m_ele_barrel_NInnerLayerLostHits_max;
+   int const    m_ele_barrel_NInnerLayerLostHits_max;
    double const m_ele_barrel_dxy_max;
    double const m_ele_barrel_swissCross_max;
    double const m_ele_barrel_r19_max;
@@ -186,7 +186,7 @@ private:
    double const m_ele_endcap_HcalD1_offset;
    double const m_ele_endcap_HcalD1_slope;
    double const m_ele_endcap_HcalD1_rhoSlope;
-   double const m_ele_endcap_NInnerLayerLostHits_max;
+   int const    m_ele_endcap_NInnerLayerLostHits_max;
    double const m_ele_endcap_dxy_max;
    double const m_ele_endcap_sigmaIetaIeta_max;
    // ID:
